+++ manager.kv
@@ -1,136 +1,5 @@
 #:kivy 1.9.0
-#:import Factory kivy.factory.Factory
 
-<CEToolBoxButton>:
-    font_size: '15sp'
-    height: '30sp'
-	#size_hint_y: None
-
-<CalculateButton@CEToolBoxButton>:
-    text: "Calculate"
-    on_release: root.on_release 
-
-<MenuButton@CEToolBoxButton>:
-    text: 'Menu'
-    on_press: root.on_press
-
-<ResetButton@CEToolBoxButton>:
-	text: 'Reset'
-	
-<CloseButton@Button>:
-	font_size: '15sp'
-    height: '30sp'
-	text: 'Close'
-	on_press: root.on_press
-	#size_hint_y: None
-
-<CEToolBoxLabel>:
-    font_size: '15sp'
-    height: '30sp'
-    width: '110sp'
-    text_size: self.width, None
-    markup: True
-
-<AboutLabel@Label>:
-	font_size: '15sp'
-	size_hint_y: None
-    text_size: self.width, None
-    height: self.texture_size[1]	
-	markup: True
-
-<CEToolBoxTextInput>:
-    font_size: '15sp'
-    height: '30sp'
-    width: '60sp'
-    #centering the text
-	padding: [6,self.height/2-10,6,self.height/2-10]
-    input_filter: 'float'
-    #waiting for a fix (floating point missing)
-    input_type: 'number'
-
-<CEToolBoxUnitLabel@Label>:
-    font_size: '15sp'
-    height: '30sp'
-    width: '30sp'
-    text_size: self.width-10, self.height
-    halign: 'left'
-    valign: 'middle'
-    #magic to use if I want a almost fixed size
-    #size_hint_y: None
-
-<CEToolBoxSpinnerOption@SpinnerOption>:
-	font_size: '15sp'
-	text_size: self.width-10, self.height
-    halign: 'left'
-    valign: 'middle'
-   
-<CEToolBoxSpinner>:
-    font_size: '15sp'
-    height: '30sp'
-	width: '30sp'
-	text_size: self.width-10, self.height
-    halign: 'left'
-    valign: 'middle'
-    option_cls: Factory.get("CEToolBoxSpinnerOption")
-	#magic to use if I want a almost fixed size
-	#size_hint_y: None
-
-<CEToolBoxLayout>:
-	padding: 10
-    spacing: 10
-	
-<TopScreenLayout>:
-    size_hint_y: None
-    size_hint_x: None
-    minimum_height: '350sp'
-    minimum_width: '350sp'
-    cols: 3
-    
-<TopPopupLayout>:
-	size_hint_y: None
-    size_hint_x: None
-    spacing: 40
-    minimum_height: '350sp'
-    minimum_width: '350sp'
-    cols: 2
-
-<CEToolBoxFixedLayout@CEToolBoxLayout>:
-	size_hint_y: None
-    height: '60sp'
-
-<CEToolBoxLabelTitle@Label>:
-	font_size: '18sp'
-	height: '20sp'
-	text_size: self.width, self.height
-	halign: 'left'
-	valign: 'middle'
-	markup: True
-
-<TitleImage@Image>:
-	source: "data/logo.png"
-	size_hint_x: None
-
-<TopTitleLayout@CEToolBoxFixedLayout>:
-	cols: 2
-	height: '50sp'
-	spacing: 5
-	pading: 5
-	
-<DownMenuLayout@CEToolBoxFixedLayout>:
-    cols: 3 
-
-<DownMenuClose@CEToolBoxFixedLayout>:
-    cols: 1
-	spacing: 0
-
-    
-<ScrollViewSpe>:
-    height: self.change_child_height(self.height)
-    width: self.change_child_width(self.width)
-
-<CEToolBoxPopup>:
-	title_size: "18sp"
-	title_align: "center"
 	
 <ErrorPopup>:
 	id: popup
